client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description =  "HTTP Streamable based example MCP Client"
requires-python = ">=3.12"
dependencies = [
    "mcp~=1.9.0",
    "python-dotenv~=1.1.0",
    "pyjwt~=2.10.1"
]